+++ custom_components/ecoflow_api/__init__.py
@@ -44,6 +44,7 @@ PLATFORMS: list[Platform] = [
     Platform.SWITCH,
     Platform.NUMBER,
     Platform.SELECT,
+    Platform.BUTTON,
 ]